+++ source/extensions/omni.isaac.lab_assets/omni/isaac/lab_assets/__init__.py
@@ -40,3 +40,4 @@ from .sawyer import *
 from .shadow_hand import *
 from .unitree import *
 from .universal_robots import *
+from .kuka import *